1. 바뀐 DTO에 맞게 회복성 패턴 테스트 코드도 수정 2. 회복성 패턴에 클래스 이름 변경 및 프로퍼티 주입 (회복성 패턴 테스트와 실제 로직 테스트는 따로 구성)
