+++ diary-service/src/main/java/com/dasd412/api/diaryservice/controller/dto/SecurityDiaryPostRequestDTO.java
@@ -17,7 +17,6 @@ import java.util.List;
 @Getter
 public class SecurityDiaryPostRequestDTO {
 
-    @NotEmpty
     @NotNull
     private final Long writerId;
