+++ diary-service/src/main/resources/application-test.properties
@@ -1,0 +1,19 @@
+
+#logging
+spring.main.banner-mode=off
+
+spring.datasource.url=jdbc:h2:~/test;
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.show_sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.jpa.hibernate.ddl-auto=create-drop
+
+spring.h2.console.path=/h2-console
+spring.h2.console.enabled=true
+
+#logging.level.org.hibernate=TRACE
+#logging.level.org.hibernate.hql=INFO
